+++ lfsapi/endpoint_finder_test.go
@@ -3,11 +3,12 @@ package lfsapi
 import (
 	"testing"
 
+	"github.com/git-lfs/git-lfs/lfshttp"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestEndpointDefaultsToOrigin(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.lfsurl": "abc",
 	}))
 
@@ -18,7 +19,7 @@ func TestEndpointDefaultsToOrigin(t *testing.T) {
 }
 
 func TestEndpointOverridesOrigin(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.url":              "abc",
 		"remote.origin.lfsurl": "def",
 	}))
@@ -30,7 +31,7 @@ func TestEndpointOverridesOrigin(t *testing.T) {
 }
 
 func TestEndpointNoOverrideDefaultRemote(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.lfsurl": "abc",
 		"remote.other.lfsurl":  "def",
 	}))
@@ -42,7 +43,7 @@ func TestEndpointNoOverrideDefaultRemote(t *testing.T) {
 }
 
 func TestEndpointUseAlternateRemote(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.lfsurl": "abc",
 		"remote.other.lfsurl":  "def",
 	}))
@@ -54,7 +55,7 @@ func TestEndpointUseAlternateRemote(t *testing.T) {
 }
 
 func TestEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "https://example.com/foo/bar",
 	}))
 
@@ -65,7 +66,7 @@ func TestEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestBareEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "https://example.com/foo/bar.git",
 	}))
 
@@ -76,7 +77,7 @@ func TestBareEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestEndpointSeparateClonePushUrl(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url":     "https://example.com/foo/bar.git",
 		"remote.origin.pushurl": "https://readwrite.com/foo/bar.git",
 	}))
@@ -93,7 +94,7 @@ func TestEndpointSeparateClonePushUrl(t *testing.T) {
 }
 
 func TestEndpointOverriddenSeparateClonePushLfsUrl(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url":        "https://example.com/foo/bar.git",
 		"remote.origin.pushurl":    "https://readwrite.com/foo/bar.git",
 		"remote.origin.lfsurl":     "https://examplelfs.com/foo/bar",
@@ -112,7 +113,7 @@ func TestEndpointOverriddenSeparateClonePushLfsUrl(t *testing.T) {
 }
 
 func TestEndpointGlobalSeparateLfsPush(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.url":     "https://readonly.com/foo/bar",
 		"lfs.pushurl": "https://write.com/foo/bar",
 	}))
@@ -129,7 +130,7 @@ func TestEndpointGlobalSeparateLfsPush(t *testing.T) {
 }
 
 func TestSSHEndpointOverridden(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url":    "git@example.com:foo/bar",
 		"remote.origin.lfsurl": "lfs",
 	}))
@@ -142,7 +143,7 @@ func TestSSHEndpointOverridden(t *testing.T) {
 }
 
 func TestSSHEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "ssh://git@example.com/foo/bar",
 	}))
 
@@ -154,7 +155,7 @@ func TestSSHEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestSSHCustomPortEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "ssh://git@example.com:9000/foo/bar",
 	}))
 
@@ -166,7 +167,7 @@ func TestSSHCustomPortEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestBareSSHEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "git@example.com:foo/bar.git",
 	}))
 
@@ -178,7 +179,7 @@ func TestBareSSHEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestBareSSSHEndpointWithCustomPortInBrackets(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "[git@example.com:2222]:foo/bar.git",
 	}))
 
@@ -190,7 +191,7 @@ func TestBareSSSHEndpointWithCustomPortInBrackets(t *testing.T) {
 }
 
 func TestSSHEndpointFromGlobalLfsUrl(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.url": "git@example.com:foo/bar.git",
 	}))
 
@@ -202,7 +203,7 @@ func TestSSHEndpointFromGlobalLfsUrl(t *testing.T) {
 }
 
 func TestHTTPEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "http://example.com/foo/bar",
 	}))
 
@@ -214,7 +215,7 @@ func TestHTTPEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestBareHTTPEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "http://example.com/foo/bar.git",
 	}))
 
@@ -226,7 +227,7 @@ func TestBareHTTPEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestGitEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "git://example.com/foo/bar",
 	}))
 
@@ -238,7 +239,7 @@ func TestGitEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestGitEndpointAddsLfsSuffixWithCustomProtocol(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "git://example.com/foo/bar",
 		"lfs.gitprotocol":   "http",
 	}))
@@ -251,7 +252,7 @@ func TestGitEndpointAddsLfsSuffixWithCustomProtocol(t *testing.T) {
 }
 
 func TestBareGitEndpointAddsLfsSuffix(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "git://example.com/foo/bar.git",
 	}))
 
@@ -263,7 +264,7 @@ func TestBareGitEndpointAddsLfsSuffix(t *testing.T) {
 }
 
 func TestLocalPathEndpointAddsDotGitDir(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "/local/path",
 	}))
 	e := finder.Endpoint("download", "")
@@ -271,7 +272,7 @@ func TestLocalPathEndpointAddsDotGitDir(t *testing.T) {
 }
 
 func TestLocalPathEndpointPreservesDotGit(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"remote.origin.url": "/local/path.git",
 	}))
 	e := finder.Endpoint("download", "")
@@ -294,7 +295,7 @@ func TestAccessConfig(t *testing.T) {
 	}
 
 	for value, expected := range tests {
-		finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+		finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 			"lfs.url":                        "http://example.com",
 			"lfs.http://example.com.access":  value,
 			"lfs.https://example.com.access": "bad",
@@ -313,7 +314,7 @@ func TestAccessConfig(t *testing.T) {
 
 	// Test again but with separate push url
 	for value, expected := range tests {
-		finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+		finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 			"lfs.url":                           "http://example.com",
 			"lfs.pushurl":                       "http://examplepush.com",
 			"lfs.http://example.com.access":     value,
@@ -340,7 +341,7 @@ func TestAccessAbsentConfig(t *testing.T) {
 }
 
 func TestSetAccess(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{}))
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{}))
 
 	assert.Equal(t, NoneAccess, finder.AccessFor("http://example.com"))
 	finder.SetAccess("http://example.com", NTLMAccess)
@@ -348,7 +349,7 @@ func TestSetAccess(t *testing.T) {
 }
 
 func TestChangeAccess(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.http://example.com.access": "basic",
 	}))
 
@@ -358,7 +359,7 @@ func TestChangeAccess(t *testing.T) {
 }
 
 func TestDeleteAccessWithNone(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.http://example.com.access": "basic",
 	}))
 
@@ -368,7 +369,7 @@ func TestDeleteAccessWithNone(t *testing.T) {
 }
 
 func TestDeleteAccessWithEmptyString(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"lfs.http://example.com.access": "basic",
 	}))
 
@@ -379,11 +380,11 @@ func TestDeleteAccessWithEmptyString(t *testing.T) {
 
 type EndpointParsingTestCase struct {
 	Given    string
-	Expected Endpoint
+	Expected lfshttp.Endpoint
 }
 
 func (c *EndpointParsingTestCase) Assert(t *testing.T) {
-	finder := NewEndpointFinder(NewContext(nil, nil, map[string]string{
+	finder := NewEndpointFinder(lfshttp.NewContext(nil, nil, map[string]string{
 		"url.https://github.com/.insteadof": "gh:",
 	}))
 	actual := finder.NewEndpoint(c.Given)
@@ -396,7 +397,7 @@ func TestEndpointParsing(t *testing.T) {
 	for desc, c := range map[string]EndpointParsingTestCase{
 		"simple bare ssh": {
 			"git@github.com:git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "https://github.com/git-lfs/git-lfs.git",
 				SshUserAndHost: "git@github.com",
 				SshPath:        "git-lfs/git-lfs.git",
@@ -406,7 +407,7 @@ func TestEndpointParsing(t *testing.T) {
 		},
 		"port bare ssh": {
 			"[git@ssh.github.com:443]:git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "https://ssh.github.com/git-lfs/git-lfs.git",
 				SshUserAndHost: "git@ssh.github.com",
 				SshPath:        "git-lfs/git-lfs.git",
@@ -416,7 +417,7 @@ func TestEndpointParsing(t *testing.T) {
 		},
 		"no user bare ssh": {
 			"github.com:git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "https://github.com/git-lfs/git-lfs.git",
 				SshUserAndHost: "github.com",
 				SshPath:        "git-lfs/git-lfs.git",
@@ -426,7 +427,7 @@ func TestEndpointParsing(t *testing.T) {
 		},
 		"bare word bare ssh": {
 			"github:git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "https://github/git-lfs/git-lfs.git",
 				SshUserAndHost: "github",
 				SshPath:        "git-lfs/git-lfs.git",
@@ -436,7 +437,7 @@ func TestEndpointParsing(t *testing.T) {
 		},
 		"insteadof alias": {
 			"gh:git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "https://github.com/git-lfs/git-lfs.git",
 				SshUserAndHost: "",
 				SshPath:        "",
@@ -446,7 +447,7 @@ func TestEndpointParsing(t *testing.T) {
 		},
 		"remote helper": {
 			"remote::git-lfs/git-lfs.git",
-			Endpoint{
+			lfshttp.Endpoint{
 				Url:            "remote::git-lfs/git-lfs.git",
 				SshUserAndHost: "",
 				SshPath:        "",
@@ -456,5 +457,23 @@ func TestEndpointParsing(t *testing.T) {
 		},
 	} {
 		t.Run(desc, c.Assert)
+	}
+}
+
+func TestNewEndpointFromCloneURLWithConfig(t *testing.T) {
+	expected := "https://foo/bar.git/info/lfs"
+	tests := []string{
+		"https://foo/bar",
+		"https://foo/bar/",
+		"https://foo/bar.git",
+		"https://foo/bar.git/",
+	}
+
+	finder := NewEndpointFinder(nil)
+	for _, actual := range tests {
+		e := finder.NewEndpointFromCloneURL(actual)
+		if e.Url != expected {
+			t.Errorf("%s returned bad endpoint url %s", actual, e.Url)
+		}
 	}
 }
