+++ lfshttp/certs_linux.go
@@ -1,4 +1,4 @@
-package lfsapi
+package lfshttp
 
 import "crypto/x509"
