+++ locking/api.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/git-lfs/git-lfs/git"
 	"github.com/git-lfs/git-lfs/lfsapi"
+	"github.com/git-lfs/git-lfs/lfshttp"
 )
 
 type lockClient struct {
@@ -56,14 +57,14 @@ func (c *lockClient) Lock(remote string, lockReq *lockRequest) (*lockResponse, *
 		return nil, nil, err
 	}
 
-	req = c.LogRequest(req, "lfs.locks.lock")
+	req = c.Client.LogRequest(req, "lfs.locks.lock")
 	res, err := c.DoWithAuth(remote, req)
 	if err != nil {
 		return nil, res, err
 	}
 
 	lockRes := &lockResponse{}
-	return lockRes, res, lfsapi.DecodeJSON(res, lockRes)
+	return lockRes, res, lfshttp.DecodeJSON(res, lockRes)
 }
 
 // UnlockRequest encapsulates the data sent in an API request to remove a lock.
@@ -101,14 +102,14 @@ func (c *lockClient) Unlock(ref *git.Ref, remote, id string, force bool) (*unloc
 		return nil, nil, err
 	}
 
-	req = c.LogRequest(req, "lfs.locks.unlock")
+	req = c.Client.LogRequest(req, "lfs.locks.unlock")
 	res, err := c.DoWithAuth(remote, req)
 	if err != nil {
 		return nil, res, err
 	}
 
 	unlockRes := &unlockResponse{}
-	err = lfsapi.DecodeJSON(res, unlockRes)
+	err = lfshttp.DecodeJSON(res, unlockRes)
 	return unlockRes, res, err
 }
 
@@ -191,7 +192,7 @@ func (c *lockClient) Search(remote string, searchReq *lockSearchRequest) (*lockL
 	}
 	req.URL.RawQuery = q.Encode()
 
-	req = c.LogRequest(req, "lfs.locks.search")
+	req = c.Client.LogRequest(req, "lfs.locks.search")
 	res, err := c.DoWithAuth(remote, req)
 	if err != nil {
 		return nil, res, err
@@ -199,7 +200,7 @@ func (c *lockClient) Search(remote string, searchReq *lockSearchRequest) (*lockL
 
 	locks := &lockList{}
 	if res.StatusCode == http.StatusOK {
-		err = lfsapi.DecodeJSON(res, locks)
+		err = lfshttp.DecodeJSON(res, locks)
 	}
 
 	return locks, res, err
@@ -251,7 +252,7 @@ func (c *lockClient) SearchVerifiable(remote string, vreq *lockVerifiableRequest
 		return nil, nil, err
 	}
 
-	req = c.LogRequest(req, "lfs.locks.verify")
+	req = c.Client.LogRequest(req, "lfs.locks.verify")
 	res, err := c.DoWithAuth(remote, req)
 	if err != nil {
 		return nil, res, err
@@ -259,7 +260,7 @@ func (c *lockClient) SearchVerifiable(remote string, vreq *lockVerifiableRequest
 
 	locks := &lockVerifiableList{}
 	if res.StatusCode == http.StatusOK {
-		err = lfsapi.DecodeJSON(res, locks)
+		err = lfshttp.DecodeJSON(res, locks)
 	}
 
 	return locks, res, err
